# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion de projet",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/email/verify/{user}": {
            "get": {
                "tags": [
                    "Inscription"
                ],
                "summary": "Vérification de l'adresse email d'un utilisateur inscrit",
                "description": "Vérification de l'adresse email d'un utilisateur inscrit",
                "operationId": "verifyEmailAddress",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Id de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 34
                        }
                    },
                    {
                        "name": "expires",
                        "in": "query",
                        "description": "Date d'expiration",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1619635176"
                        }
                    },
                    {
                        "name": "signature",
                        "in": "query",
                        "description": "Signature",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "a0dbd2d1bb02980b25881132262dae0ae9780d9c51a2c7fe79359badeb0dd508"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vérification réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Message de réussite de la vérification",
                                            "type": "string",
                                            "example": "Votre adresse email est vérifiée."
                                        },
                                        "user": {
                                            "description": "Détails de l'utilisateur",
                                            "properties": {
                                                "id": {
                                                    "description": "ID",
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "name": {
                                                    "description": "Nom d'usager",
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "firstname": {
                                                    "description": "Prénom(s)",
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "lastname": {
                                                    "description": "Nom",
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "email": {
                                                    "description": "Email",
                                                    "type": "string",
                                                    "example": "mathieu.roy@strategemedia.com"
                                                },
                                                "company": {
                                                    "description": "Compagnie",
                                                    "type": "string",
                                                    "example": "Alu-Rex"
                                                },
                                                "mobilephone": {
                                                    "description": "Cellulaire",
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "phone": {
                                                    "description": "Téléphone",
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "email_verified_at": {
                                                    "description": "Date de vérification de l'email",
                                                    "type": "string",
                                                    "example": "2021-04-28T21:28:53.000000Z"
                                                },
                                                "created_at": {
                                                    "description": "Date de création",
                                                    "type": "string",
                                                    "example": "2021-04-28T21:21:35.000000Z"
                                                },
                                                "updated_at": {
                                                    "description": "Date de mis à jour",
                                                    "type": "string",
                                                    "example": "2021-04-28T21:28:53.000000Z"
                                                },
                                                "deleted_at": {
                                                    "description": "Date de suppression",
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "stripe_id": {
                                                    "description": "Identifiant Client de Stripe",
                                                    "type": "string",
                                                    "example": "cus_JTGwkBQuLY8Nqy"
                                                },
                                                "card_brand": {
                                                    "description": "Type de carte de paiement",
                                                    "type": "string",
                                                    "example": "mastercard"
                                                },
                                                "card_last_four": {
                                                    "description": "Les 4 derniers chiffres de la carte",
                                                    "type": "string",
                                                    "example": "4444"
                                                },
                                                "trial_ends_at": {
                                                    "description": "Date de fin de l'essai",
                                                    "type": "string",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Vérification échouée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Message d'échec de vérification",
                                            "type": "string",
                                            "example": "La vérification de votre adresse email a échouée."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Adresse email déjà vérifiée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Vérification déjà faite",
                                            "type": "string",
                                            "example": "Votre adresse email a déjà été vérifiée."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "Sanctum"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}